Pull Renderer into nonrealtime.sessions
